image.pollinations.ai/nunchaku/requirements.txt
# Flux Schnell Server Dependencies
# Note: PyTorch is installed separately with CUDA support via setup.sh

# Core ML
accelerate>=1.9
diffusers>=0.35
transformers>=4.50
safetensors
sentencepiece
protobuf
peft
einops

# Server
fastapi
uvicorn
pydantic
python-multipart
aiohttp

# Utils
pillow
numpy
tqdm
requests
huggingface-hub
python-dotenv
